# docs/swagger.yaml
definitions:
  domain.Bond:
    properties:
      couponRate:
        type: number
      id:
        type: string
      issuer:
        type: string
      maturityDate:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
  domain.CMSP:
    properties:
      description:
        type: string
      id:
        type: integer
      licensedDate:
        type: string
      name:
        type: string
      plan:
        type: string
      source:
        type: string
      type:
        type: string
    type: object
  domain.Stock:
    properties:
      changePercent:
        type: number
      name:
        type: string
      price:
        type: number
      symbol:
        type: string
      volume:
        type: integer
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/bonds/search:
    get:
      consumes:
      - application/json
      description: Search bonds by coupon rate and maturity date
      parameters:
      - description: Minimum coupon rate
        in: query
        name: minCoupon
        type: number
      - description: Maximum coupon rate
        in: query
        name: maxCoupon
        type: number
      - description: Maturity before (YYYY-MM-DD)
        in: query
        name: maturityBefore
        type: string
      - description: Maturity after (YYYY-MM-DD)
        in: query
        name: maturityAfter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Bond'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      summary: Search bonds
      tags:
      - bonds
  /api/v1/cmsp:
    get:
      consumes:
      - application/json
      description: Get all Capital Market Service Providers, filter by type or license
        date
      parameters:
      - description: CMSP type
        in: query
        name: type
        type: string
      - description: Licensed before (YYYY-MM-DD)
        in: query
        name: licensedBefore
        type: string
      - description: Licensed after (YYYY-MM-DD)
        in: query
        name: licensedAfter
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CMSP'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      summary: List all CMSPs
      tags:
      - cmsp
  /api/v1/cmsp/{id}:
    get:
      consumes:
      - application/json
      description: Get a Capital Market Service Provider by its ID
      parameters:
      - description: CMSP ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CMSP'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      summary: Get CMSP by ID
      tags:
      - cmsp
  /api/v1/stocks:
    get:
      consumes:
      - application/json
      description: Get all stocks, optionally filter by gainer or loser
      parameters:
      - description: gainer or loser
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Stock'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      summary: List all stocks
      tags:
      - stocks
swagger: "2.0"